MultiView-Cube/flow/tmpl_measure/requirements.txt
pytz
PyYAML
kafka-python
uptime